149 - Cleanup query mappers from generated types
